добавлен метод update; добавлена проверка в методах на наличие передаваемого резюме; добавлена в save проверка на преполнение устранено дублирование;
