--- src/Resume.java
@@ -1,13 +1,0 @@
-/**
- * com.urise.webapp.model.Resume class
- */
-public class Resume {
-
-    // Unique identifier
-    String uuid;
-
-    @Override
-    public String toString() {
-        return uuid;
-    }
-}
