+++ src/com/urise/webapp/model/Resume.java
@@ -1,0 +1,23 @@
+package com.urise.webapp.model;
+
+/**
+ * com.urise.webapp.model.com.urise.webapp.model.Resume class
+ */
+public class Resume {
+
+    // Unique identifier
+    private String uuid;
+
+    @Override
+    public String toString() {
+        return uuid;
+    }
+
+    public String getUuid() {
+        return uuid;
+    }
+
+    public void setUuid(String uuid) {
+        this.uuid = uuid;
+    }
+}
